Added Chance module as an alternative to Math.random() for entropy
Math.random() is a low entropy source for randomness and also since
it has limited resolution, sometimes it generates things with lots of
bias.

Chance is a Mersenne Twister implementation, so when initialized for
X and Y and if the ship is oriented on X or Y axis using Math.random()
as a source of entropy, we have now 3 Mersenne Twisters, all accessed
using different methods.

The improvement is that for example, Chance when asking for a Boolean,
has actual 50% chance of either giving true or false, so for 5 calls we
never should get a series of the same values, as the output is a bit
"biased" towards expected randomness, not real randomness.

Also, when testing, we can provide our own IVs for Mersenne Twister,
which guarantee reproducibility of the results.

Reorganized the code a bit, Battleground now has BattlegroundMap,
ready to implement GameLogic.